fix: address CodeRabbit review feedback on PR #557
- Bump pypdf>=4.0.0 to >=6.8.0 (excludes DoS CVEs)
- Remove unused init_side_effect and _mock_download dead code
- Extract duplicated _make_minimal_pdf to shared tests/helpers.py
- Move io import to module level in test_pdf_image_utils.py
